+++ ingestion_tools/scripts/standardize_dirs.py
@@ -16,6 +16,8 @@ def cli(ctx):
 
 def common_options(func):
     options = []
+    options.append(click.option("--import-everything", is_flag=True, default=False))
+    options.append(click.option("--import-all-metadata", is_flag=True, default=False))
     for cls in IMPORTERS:
         plural_key = cls.plural_key.replace("_", "-")
         importer_key = cls.type_key.replace("_", "-")
@@ -80,7 +82,7 @@ def do_import(config, tree, to_import, metadata_import, to_iterate, kwargs, pare
 @click.argument("config_file", required=True, type=str)
 @click.argument("input_bucket", required=True, type=str)
 @click.argument("output_path", required=True, type=str)
-@click.option("--import-everything", is_flag=True, default=False)
+@click.option("--https-prefix", required=False, type=str, help="protocol + domain for where to fetch files via HTTP")
 @click.option("--write-mrc/--no-write-mrc", default=True)
 @click.option("--write-zarr/--no-write-zarr", default=True)
 @click.option("--force-overwrite", is_flag=True, default=False)
@@ -92,7 +94,9 @@ def convert(
     config_file: str,
     input_bucket: str,
     output_path: str,
+    https_prefix: str,
     import_everything: bool,
+    import_all_metadata: bool,
     write_mrc: bool,
     write_zarr: bool,
     force_overwrite: bool,
@@ -105,7 +109,7 @@ def convert(
 
     fs = FileSystemApi.get_fs_api(mode=fs_mode, force_overwrite=force_overwrite)
 
-    config = DepositionImportConfig(fs, config_file, output_path, input_bucket, IMPORTERS)
+    config = DepositionImportConfig(fs, config_file, output_path, input_bucket, IMPORTERS, https_prefix=https_prefix)
     config.write_mrc = write_mrc
     config.write_zarr = write_zarr
     config.load_map_files()
@@ -113,6 +117,10 @@ def convert(
     iteration_deps = flatten_dependency_tree(IMPORTER_DEP_TREE).items()
     if import_everything:
         to_import = set(IMPORTERS)
+        metadata_import = set(IMPORTERS)
+        to_iterate = set(IMPORTERS)
+    elif import_all_metadata:
+        to_import = set()
         metadata_import = set(IMPORTERS)
         to_iterate = set(IMPORTERS)
     else: