+++ schema/core/v2.0.0/codegen/metadata_models.py
@@ -135,6 +135,20 @@ linkml_meta = LinkMLMeta(
                 "name": "EMPIAR_ID",
                 "pattern": "^EMPIAR-[0-9]+$",
             },
+            "FBBT_ID": {
+                "base": "str",
+                "description": "A drosophila anatomy ontology " "identifier",
+                "from_schema": "metadata",
+                "name": "FBBT_ID",
+                "pattern": "FBbt:[0-9]{8}$",
+            },
+            "FBDV_ID": {
+                "base": "str",
+                "description": "A drosophila developmental stage " "ontology identifier",
+                "from_schema": "metadata",
+                "name": "FBDV_ID",
+                "pattern": "FBdv:[0-9]{8}$",
+            },
             "FloatFormattedString": {
                 "base": "str",
                 "description": "A formatted string that " "represents a floating " "point number.",
@@ -151,7 +165,7 @@ linkml_meta = LinkMLMeta(
             },
             "HSAPDV_ID": {
                 "base": "str",
-                "description": "A Human Developmental Phenotype " "Ontology identifier",
+                "description": "A human developmental phenotype " "ontology identifier",
                 "from_schema": "metadata",
                 "name": "HSAPDV_ID",
                 "pattern": "HsapDv:[0-9]{7}$",
@@ -162,6 +176,13 @@ linkml_meta = LinkMLMeta(
                 "from_schema": "metadata",
                 "name": "IntegerFormattedString",
                 "pattern": "^int[ " "]*\\{[a-zA-Z0-9_-]+\\}[ ]*$",
+            },
+            "MMUSDV_ID": {
+                "base": "str",
+                "description": "A mouse developmental stage ontology " "identifier",
+                "from_schema": "metadata",
+                "name": "MMUSDV_ID",
+                "pattern": "MmusDv:[0-9]{7}$",
             },
             "MONDO_ID": {
                 "base": "str",
@@ -247,9 +268,16 @@ linkml_meta = LinkMLMeta(
                 "minimum_value": 0,
                 "name": "VersionString",
             },
+            "WBBT_ID": {
+                "base": "str",
+                "description": "A WormBase anatomy ontology identifier",
+                "from_schema": "metadata",
+                "name": "WBBT_ID",
+                "pattern": "WBbt:[0-9]{7}$",
+            },
             "WORMBASE_DEVELOPMENT_ID": {
                 "base": "str",
-                "description": "A WormBase identifier",
+                "description": "A WormBase " "developmental stage " "identifier",
                 "from_schema": "metadata",
                 "name": "WORMBASE_DEVELOPMENT_ID",
                 "pattern": "WBls:[0-9]{7}$",
@@ -261,12 +289,19 @@ linkml_meta = LinkMLMeta(
                 "name": "WORMBASE_STRAIN_ID",
                 "pattern": "WBStrain[0-9]{8}$",
             },
-            "WORMBASE_TISSUE_ID": {
+            "ZFA_ID": {
                 "base": "str",
-                "description": "A WormBase tissue identifier",
+                "description": "A zebrafish anatomy ontology identifier",
                 "from_schema": "metadata",
-                "name": "WORMBASE_TISSUE_ID",
-                "pattern": "WBbt:[0-9]{7}$",
+                "name": "ZFA_ID",
+                "pattern": "ZFA:[0-9]{7}$",
+            },
+            "ZFS_ID": {
+                "base": "str",
+                "description": "A zebrafish developmental stage ontology " "identifier",
+                "from_schema": "metadata",
+                "name": "ZFS_ID",
+                "pattern": "ZFS:[0-9]{7}$",
             },
             "boolean": {
                 "base": "Bool",
@@ -683,6 +718,8 @@ class CtfFormatEnum(str, Enum):
 
     # The file has ctffind schema
     CTFFIND = "CTFFIND"
+    # The file has Gctf schema
+    Gctf = "Gctf"
 
 
 class DepositionTypesEnum(str, Enum):
@@ -1166,6 +1203,9 @@ class DevelopmentStageDetails(ConfiguredBaseModel):
                     {"range": "WORMBASE_DEVELOPMENT_ID"},
                     {"range": "UBERON_ID"},
                     {"range": "HSAPDV_ID"},
+                    {"range": "MMUSDV_ID"},
+                    {"range": "ZFS_ID"},
+                    {"range": "FBDV_ID"},
                 ],
                 "domain_of": [
                     "Assay",
@@ -1185,7 +1225,9 @@ class DevelopmentStageDetails(ConfiguredBaseModel):
 
     @field_validator("id")
     def pattern_id(cls, v):
-        pattern = re.compile(r"(^unknown$)|(WBls:[0-9]{7}$)|(^UBERON:[0-9]{7}$)|(HsapDv:[0-9]{7}$)")
+        pattern = re.compile(
+            r"(^unknown$)|(WBls:[0-9]{7}$)|(^UBERON:[0-9]{7}$)|(HsapDv:[0-9]{7}$)|(MmusDv:[0-9]{7}$)|(ZFS:[0-9]{7}$)|(FBdv:[0-9]{8}$)"
+        )
         if isinstance(v, list):
             for element in v:
                 if not pattern.match(element):
@@ -1347,7 +1389,13 @@ class TissueDetails(ConfiguredBaseModel):
         json_schema_extra={
             "linkml_meta": {
                 "alias": "id",
-                "any_of": [{"range": "CL_ID"}, {"range": "WORMBASE_TISSUE_ID"}, {"range": "UBERON_ID"}],
+                "any_of": [
+                    {"range": "CL_ID"},
+                    {"range": "WBBT_ID"},
+                    {"range": "ZFA_ID"},
+                    {"range": "FBBT_ID"},
+                    {"range": "UBERON_ID"},
+                ],
                 "domain_of": [
                     "Assay",
                     "DevelopmentStageDetails",
@@ -1366,7 +1414,7 @@ class TissueDetails(ConfiguredBaseModel):
 
     @field_validator("id")
     def pattern_id(cls, v):
-        pattern = re.compile(r"(^CL:[0-9]{7}$)|(WBbt:[0-9]{7}$)|(^UBERON:[0-9]{7}$)")
+        pattern = re.compile(r"(^CL:[0-9]{7}$)|(WBbt:[0-9]{7}$)|(ZFA:[0-9]{7}$)|(FBbt:[0-9]{8}$)|(^UBERON:[0-9]{7}$)")
         if isinstance(v, list):
             for element in v:
                 if not pattern.match(element):
@@ -1414,7 +1462,13 @@ class CellType(ConfiguredBaseModel):
         json_schema_extra={
             "linkml_meta": {
                 "alias": "id",
-                "any_of": [{"range": "CL_ID"}, {"range": "UBERON_ID"}],
+                "any_of": [
+                    {"range": "CL_ID"},
+                    {"range": "WBBT_ID"},
+                    {"range": "ZFA_ID"},
+                    {"range": "FBBT_ID"},
+                    {"range": "UBERON_ID"},
+                ],
                 "domain_of": [
                     "Assay",
                     "DevelopmentStageDetails",
@@ -1433,7 +1487,7 @@ class CellType(ConfiguredBaseModel):
 
     @field_validator("id")
     def pattern_id(cls, v):
-        pattern = re.compile(r"(^CL:[0-9]{7}$)|(^UBERON:[0-9]{7}$)")
+        pattern = re.compile(r"(^CL:[0-9]{7}$)|(WBbt:[0-9]{7}$)|(ZFA:[0-9]{7}$)|(FBbt:[0-9]{8}$)|(^UBERON:[0-9]{7}$)")
         if isinstance(v, list):
             for element in v:
                 if not pattern.match(element):
@@ -5410,7 +5464,7 @@ class Ctf(ConfiguredBaseModel):
 
     @field_validator("format")
     def pattern_format(cls, v):
-        pattern = re.compile(r"^CTFFIND$")
+        pattern = re.compile(r"(^CTFFIND$)|(^Gctf$)")
         if isinstance(v, list):
             for element in v:
                 if not pattern.match(element):